test: add instrumentation testing utilities and cleanup fixture
